Revert "Merge pull request #979 from MmKolodziej/better_total_count_for_grouped_queries"
This reverts commit b5e2e07234caf015ca802e647a44512ebe45244a, reversing
changes made to d1c26fd2fcd16f29b00bb076f47445ccbc550b89.

We merged #979 as a performance improvement, but it turns out it
introduced a couple of bugs. The solution seems to be a little tricky as
it has to call `.unscoped` and `unscope(where: :type)`, so let's revert
it until we find a better solution.

closes #1012, #1015.